Multimodel OCR/requirements.txt
transformers
gradio
Pillow
torch
